Add input mappers for anvil-catalog entities (#600)
