+++ explorer/files/anvil-catalog/build-workspaces.ts
@@ -1,8 +1,37 @@
-import {
-  AnVILCatalog,
-  AnVILCatalogWorkspace,
-} from "../../app/apis/catalog/anvil-catalog/common/entities";
+import { AnVILCatalogWorkspace } from "../../app/apis/catalog/anvil-catalog/common/entities";
 import { DbGapStudy, getStudy } from "../common/dbGaP";
+
+export interface AnVILCatalog {
+  bucketName: string;
+  bucketSize: number;
+  COL: string;
+  consentLongName: string;
+  consentTitle: string;
+  consortium: string;
+  discoveryCount: number;
+  diseaseText: string;
+  DS: string;
+  familyCount: number;
+  GRU: string;
+  GSO: string;
+  HMB: string;
+  IRB: string;
+  "library:datatype": string[];
+  "library:dataUseRestriction": string;
+  "library:indication": string[];
+  "library:studyDesign": string[];
+  MDS: string;
+  name: string;
+  NPU: string;
+  NRES: string;
+  participantCount: number;
+  phsId: string;
+  PUB: string;
+  requestorPays: boolean;
+  sampleCount: number;
+  status: string;
+  subjectCount: number;
+}
 
 /**
  * Returns AnVIL catalog workspaces.
